obexd: Fix free space calculation on systems where fsfilcnt_t is 32-bit
On some platforms f_bavail will be a 32-bit variable and in that case we
need to make sure that the multiplication of it with f_bsize is done in
64-bits. Therefore, explicitly typecast f_bsize to 64-bit before the
multiplication.

+++ obexd/src/obex.c
@@ -751,7 +751,7 @@ static gboolean check_put(obex_t *obex, obex_object_t *obj)
 		return FALSE;
 	}
 
-	free = buf.f_bsize * buf.f_bavail;
+	free = (unsigned long long) buf.f_bsize * buf.f_bavail;
 	debug("Free space in disk: %llu", free);
 	if ((guint64) os->size > free) {
 		debug("Free disk space not available");
